+++ src/basic/time.h
@@ -1,8 +1,8 @@
 void wait_msec(unsigned int n);
 
-void start_timer(int *start_time);
+void start_timer(unsigned long *start_time);
 
-int elapsed_time(int start_time);
+int elapsed_time(unsigned long start_time);
 
 int uptime_msec();
